refactor(sfn): unify the duplicate JSONPath resolvers
